// app/types/gameEnums.ts
export enum solutionTypes {
  THREADED = "threaded",
  SEQUENTIAL = "sequential",
  COMPARISON = "comparison",
  GAME_PLAY = "game play",
}

export enum gameStatusType {
  WIN = "win",
  LOST = "lost",
  DRAW = "draw",
}

export enum SolutionRecognition {
  Unique = "unique",
  Known = "known",
  Variation = "variation",
}